+++ packages/youtube-mute-skip-ads/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add a missing `if (debugging)` around a `console.debug` call.
   - TODO: Add logging functions which handle the prefix and the debug mode.
+- Add a build with debug logging.
 
 ## 0.0.27 2024-09-25